+++ zephyr-mcumgr/src/client.rs
@@ -194,7 +194,7 @@ impl MCUmgrClient {
     /// # use zephyr_mcumgr::MCUmgrClient;
     /// # fn main() {
     /// let serial = serialport::new("COM42", 115200)
-    ///     .timeout(std::time::Duration::from_millis(500))
+    ///     .timeout(std::time::Duration::from_millis(2000))
     ///     .open()
     ///     .unwrap();
     ///
@@ -484,6 +484,18 @@ impl MCUmgrClient {
         self.connection
             .execute_command(&commands::image::GetImageState)
             .map(|val| val.images)
+    }
+
+    /// Erase image slot on target device.
+    ///
+    /// # Arguments
+    ///
+    /// * `slot` - The slot ID of the image to erase. Slot `1` if omitted.
+    ///
+    pub fn image_erase(&self, slot: Option<u32>) -> Result<(), ExecuteError> {
+        self.connection
+            .execute_command(&commands::image::ImageErase { slot })
+            .map(Into::into)
     }
 
     /// Obtain a list of available image slots.