+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.6.2] - 2026-01-13
+
+### Changes
+
+- Add Python/Rust library commands:
+  - `image_erase`
+- Add CLI commands:
+  - `image`
+    - `erase`
+- Increase default timeout to `2000 ms`.
+
 ## [0.6.1] - 2026-01-13
 
 ### Changes
@@ -201,6 +212,7 @@ Initial release, not feature complete yet.
 
 Primarily to test release workflow.
 
+[0.6.2]: https://github.com/Finomnis/zephyr-mcumgr-client/compare/0.6.1...0.6.2
 [0.6.1]: https://github.com/Finomnis/zephyr-mcumgr-client/compare/0.6.0...0.6.1
 [0.6.0]: https://github.com/Finomnis/zephyr-mcumgr-client/compare/0.5.1...0.6.0
 [0.5.1]: https://github.com/Finomnis/zephyr-mcumgr-client/compare/0.5.0...0.5.1